refactor: small touchups, update util imports
