+++ src/components/box.ts
@@ -1,14 +1,12 @@
 import { Component, ComponentOptions } from "../component.ts";
 import { Rectangle } from "../types.ts";
-import { EventRecord } from "../util.ts";
+import { EventRecord } from "../utils/typed_event_target.ts";
 
 export interface BoxComponentOptions extends ComponentOptions {
   rectangle: Rectangle;
 }
 
-export class BoxComponent<
-  EventMap extends EventRecord = Record<never, never>,
-> extends Component<EventMap> {
+export class BoxComponent<EventMap extends EventRecord = Record<never, never>> extends Component<EventMap> {
   declare rectangle: Rectangle;
 
   constructor(options: BoxComponentOptions) {
