+++ src/components/label.ts
@@ -1,6 +1,7 @@
 import { Component, ComponentOptions } from "../component.ts";
 import { Rectangle } from "../types.ts";
-import { EventRecord, textWidth } from "../util.ts";
+import { textWidth } from "../utils/strings.ts";
+import { EventRecord } from "../utils/typed_event_target.ts";
 
 export interface LabelTextAlign {
   horizontal: "left" | "center" | "right";
@@ -13,9 +14,7 @@ export interface LabelComponentOptions extends ComponentOptions {
   align: LabelTextAlign;
 }
 
-export class LabelComponent<
-  EventMap extends EventRecord = Record<never, never>,
-> extends Component<EventMap> {
+export class LabelComponent<EventMap extends EventRecord = Record<never, never>> extends Component<EventMap> {
   declare rectangle: Rectangle;
   value: string;
   align: LabelTextAlign;
